P-value calculation threshold for junctions implemented

+++ src/c/breseq/libbreseq/chisquare.h
@@ -9,9 +9,9 @@ namespace breseq {
 	double incompletegamma(double a, double x, bool complemented = false);
 	double lngamma(double x, double* sgngam);
   
-  double nbdtrc(int k, int n, double p);
-  double nbdtr(int k, int n, double p);
-  double nbdtri(int k, int n, double p);
+  double nbdtrc(int k, double n, double p);
+  double nbdtr(int k, double n, double p);
+  double nbdtri(int k, double n, double p);
   
   double incbi(double aa, double bb, double yy0);
   double incbet(double aa, double bb, double xx);
